+++ src/contractModels/FalseContract.java
@@ -2,7 +2,7 @@ package contractModels;
 
 public class FalseContract extends Contract {
 	@Override
-	public int timeout() {
+	public long timeout() {
 		return Constants.pInf;
 	}
 }
